+++ lib/promise-socket.d.ts
@@ -1,0 +1,15 @@
+import * as net from 'net'
+
+import { PromiseDuplex } from 'promise-duplex'
+
+export class PromiseSocket<TSocket extends net.Socket> extends PromiseDuplex<TSocket> {
+  constructor (socket?: TSocket)
+
+  connect(port: number): Promise<void>
+  connect(path: string): Promise<void>
+  connect(options: object): Promise<void>
+
+  setTimeout (timeout: number): Promise<void>
+}
+
+export default PromiseSocket