+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.0.0 2017-10-02
+
+  * Exports also as a class and namespace and the default.
+  * Typings for Typescript.
+  * Based on promise-duplex@1.x.x
+
 ## v0.0.2 2017-06-22
 
   * Upgraded chai@4.0.2, chai-as-promised@7.0.0, promise-duplex@0.1.2,